Automate tagging/version-bump
Much like with Drone, we perform the version bump with a commit message that tells the CI system to ignore *that* commit, so that we don't get recursion ad infinitum.  The version bump and publish steps are filtered so that they only apply to the `master` branch (so they don't happen on PR builds).

Updated CI build to use node 10.6 in order to get `npm ci`, which obsoletes the `restore_cache`/`save_cache` CircleCI pattern.

Added `persist_to_workspace`/`attach_workspace` to get common context for the duration of the workflow.